# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vizitka API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  fullName: string;
  /** @nullable */
  title?: string | null;
  bio: string;
  phone: string;
  instagram: string;
  telegram: string;
  /** @nullable */
  photoUrl: string | null;
  createdAt?: string;
}

export interface ProfileInput {
  /** @minLength 1 */
  fullName: string;
  /** @nullable */
  title?: string | null;
  bio: string;
  phone: string;
  instagram: string;
  telegram: string;
  /** @nullable */
  photoUrl?: string | null;
}

export interface AdminCredentials {
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}