# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vizitka API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: profile
    description: Profile operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get profile
      description: Returns the public profile information
      responses:
        "200":
          description: Profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile
      description: Update the profile (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [profile]
      summary: Admin login
      description: Login to the admin panel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [profile]
      summary: Admin logout
      description: Logout from admin panel
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [profile]
      summary: Check admin session
      description: Check if current session is authenticated admin
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Profile:
      type: object
      required: [id, fullName, bio, phone, instagram, telegram, photoUrl]
      properties:
        id:
          type: integer
        fullName:
          type: string
        title:
          type: ["string", "null"]
        bio:
          type: string
        phone:
          type: string
        instagram:
          type: string
        telegram:
          type: string
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    ProfileInput:
      type: object
      required: [fullName, bio, phone, instagram, telegram]
      properties:
        fullName:
          type: string
          minLength: 1
        title:
          type: ["string", "null"]
        bio:
          type: string
        phone:
          type: string
        instagram:
          type: string
        telegram:
          type: string
        photoUrl:
          type: ["string", "null"]

    AdminCredentials:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AdminLoginResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string