# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vizitka API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  fullName: string;
  /** @nullable */
  title?: string | null;
  bio: string;
  phone: string;
  instagram: string;
  telegram: string;
  /** @nullable */
  photoUrl: string | null;
  createdAt?: string;
}